Refractoring
Took out code that adds graphs to div and moved to separate file.
